Split model_clients into 3 files
