+++ lumigator/jobs/inference/model_clients/base_client.py
@@ -1,0 +1,23 @@
+from abc import abstractmethod
+
+from inference_config import InferenceJobConfig
+
+from schemas import PredictionResult
+
+
+class BaseModelClient:
+    """Abstract class for a model client, used to provide a uniform interface
+    (currentnly just a simple predict method) to models served in different
+    ways (e.g. HF models loaded locally, OpenAI endpoints, vLLM inference
+    servers, llamafile).
+    """
+
+    @abstractmethod
+    def __init__(self, model: str, config: InferenceJobConfig):
+        """Used to initialize the model / inference service."""
+        pass
+
+    @abstractmethod
+    def predict(self, prompt: str) -> PredictionResult:
+        """Given a prompt, return a prediction."""
+        pass
